koji_target: Do not delete target in check_mode
Fixes: https://github.com/ktdreyer/koji-ansible/issues/130

+++ library/koji_target.py
@@ -99,10 +99,11 @@ def delete_target(session, name, check_mode):
         changed=False,
     )
     if targetinfo:
-        common_koji.ensure_logged_in(session)
-        session.deleteBuildTarget(targetinfo['id'])
         result['stdout'] = 'deleted target %d' % targetinfo['id']
         result['changed'] = True
+        if not check_mode:
+            common_koji.ensure_logged_in(session)
+            session.deleteBuildTarget(targetinfo['id'])
     return result
